+++ lib/features/offeredfood/domain/model/food_info.freezed.dart
@@ -16,12 +16,19 @@ final _privateConstructorUsedError = UnsupportedError(
 
 /// @nodoc
 mixin _$FoodInfo {
+  String get id => throw _privateConstructorUsedError; // The UUID identifier
   String? get dishName => throw _privateConstructorUsedError;
+
   List<String>? get allergens => throw _privateConstructorUsedError;
+
   String? get foodCategory => throw _privateConstructorUsedError;
+
   int? get numberOfServings => throw _privateConstructorUsedError;
+
   int? get numberOfBoxes => throw _privateConstructorUsedError;
+
   String? get foodBoxId => throw _privateConstructorUsedError;
+
   DateTime? get consumeBy => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
@@ -33,9 +40,11 @@ mixin _$FoodInfo {
 abstract class $FoodInfoCopyWith<$Res> {
   factory $FoodInfoCopyWith(FoodInfo value, $Res Function(FoodInfo) then) =
       _$FoodInfoCopyWithImpl<$Res, FoodInfo>;
+
   @useResult
   $Res call(
-      {String? dishName,
+      {String id,
+      String? dishName,
       List<String>? allergens,
       String? foodCategory,
       int? numberOfServings,
@@ -51,12 +60,14 @@ class _$FoodInfoCopyWithImpl<$Res, $Val extends FoodInfo>
 
   // ignore: unused_field
   final $Val _value;
+
   // ignore: unused_field
   final $Res Function($Val) _then;
 
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? id = null,
     Object? dishName = freezed,
     Object? allergens = freezed,
     Object? foodCategory = freezed,
@@ -66,6 +77,10 @@ class _$FoodInfoCopyWithImpl<$Res, $Val extends FoodInfo>
     Object? consumeBy = freezed,
   }) {
     return _then(_value.copyWith(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as String,
       dishName: freezed == dishName
           ? _value.dishName
           : dishName // ignore: cast_nullable_to_non_nullable
@@ -104,10 +119,12 @@ abstract class _$$FoodInfoImplCopyWith<$Res>
   factory _$$FoodInfoImplCopyWith(
           _$FoodInfoImpl value, $Res Function(_$FoodInfoImpl) then) =
       __$$FoodInfoImplCopyWithImpl<$Res>;
+
   @override
   @useResult
   $Res call(
-      {String? dishName,
+      {String id,
+      String? dishName,
       List<String>? allergens,
       String? foodCategory,
       int? numberOfServings,
@@ -127,6 +144,7 @@ class __$$FoodInfoImplCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? id = null,
     Object? dishName = freezed,
     Object? allergens = freezed,
     Object? foodCategory = freezed,
@@ -136,6 +154,10 @@ class __$$FoodInfoImplCopyWithImpl<$Res>
     Object? consumeBy = freezed,
   }) {
     return _then(_$FoodInfoImpl(
+      id: null == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as String,
       dishName: freezed == dishName
           ? _value.dishName
           : dishName // ignore: cast_nullable_to_non_nullable
@@ -172,7 +194,8 @@ class __$$FoodInfoImplCopyWithImpl<$Res>
 
 class _$FoodInfoImpl implements _FoodInfo {
   const _$FoodInfoImpl(
-      {this.dishName,
+      {required this.id,
+      this.dishName,
       final List<String>? allergens,
       this.foodCategory,
       this.numberOfServings,
@@ -182,8 +205,13 @@ class _$FoodInfoImpl implements _FoodInfo {
       : _allergens = allergens;
 
   @override
+  final String id;
+
+// The UUID identifier
+  @override
   final String? dishName;
   final List<String>? _allergens;
+
   @override
   List<String>? get allergens {
     final value = _allergens;
@@ -206,7 +234,7 @@ class _$FoodInfoImpl implements _FoodInfo {
 
   @override
   String toString() {
-    return 'FoodInfo(dishName: $dishName, allergens: $allergens, foodCategory: $foodCategory, numberOfServings: $numberOfServings, numberOfBoxes: $numberOfBoxes, foodBoxId: $foodBoxId, consumeBy: $consumeBy)';
+    return 'FoodInfo(id: $id, dishName: $dishName, allergens: $allergens, foodCategory: $foodCategory, numberOfServings: $numberOfServings, numberOfBoxes: $numberOfBoxes, foodBoxId: $foodBoxId, consumeBy: $consumeBy)';
   }
 
   @override
@@ -214,6 +242,7 @@ class _$FoodInfoImpl implements _FoodInfo {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$FoodInfoImpl &&
+            (identical(other.id, id) || other.id == id) &&
             (identical(other.dishName, dishName) ||
                 other.dishName == dishName) &&
             const DeepCollectionEquality()
@@ -233,6 +262,7 @@ class _$FoodInfoImpl implements _FoodInfo {
   @override
   int get hashCode => Object.hash(
       runtimeType,
+      id,
       dishName,
       const DeepCollectionEquality().hash(_allergens),
       foodCategory,
@@ -250,7 +280,8 @@ class _$FoodInfoImpl implements _FoodInfo {
 
 abstract class _FoodInfo implements FoodInfo {
   const factory _FoodInfo(
-      {final String? dishName,
+      {required final String id,
+      final String? dishName,
       final List<String>? allergens,
       final String? foodCategory,
       final int? numberOfServings,
@@ -259,19 +290,29 @@ abstract class _FoodInfo implements FoodInfo {
       final DateTime? consumeBy}) = _$FoodInfoImpl;
 
   @override
+  String get id;
+
+  @override
   String? get dishName;
+
   @override
   List<String>? get allergens;
+
   @override
   String? get foodCategory;
+
   @override
   int? get numberOfServings;
+
   @override
   int? get numberOfBoxes;
+
   @override
   String? get foodBoxId;
+
   @override
   DateTime? get consumeBy;
+
   @override
   @JsonKey(ignore: true)
   _$$FoodInfoImplCopyWith<_$FoodInfoImpl> get copyWith =>