+++ lib/models/dto/meal_dto.g.dart
@@ -8,11 +8,11 @@ part of 'meal_dto.dart';
 
 MealDto _$MealDtoFromJson(Map<String, dynamic> json) => MealDto(
       mealId: json['mealId'] as String,
-      count: json['count'] as int,
+      count: (json['count'] as num).toInt(),
       consumeBy:
           const TimestampConverter().fromJson(json['consumeBy'] as Timestamp),
       foodBoxId: json['foodBoxId'] as String,
-      foodBoxCount: json['foodBoxCount'] as int,
+      foodBoxCount: (json['foodBoxCount'] as num).toInt(),
     );
 
 Map<String, dynamic> _$MealDtoToJson(MealDto instance) => <String, dynamic>{