+++ lib/routes/app_router.gr.dart
@@ -8,13 +8,13 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:auto_route/auto_route.dart' as _i14;
-import 'package:flutter/material.dart' as _i15;
+import 'package:auto_route/auto_route.dart' as _i16;
+import 'package:flutter/material.dart' as _i17;
 import 'package:zachranobed/features/appTerms/presentation/app_terms_screen.dart'
     as _i1;
 import 'package:zachranobed/features/debug/debug_screen.dart' as _i5;
 import 'package:zachranobed/features/foodboxes/domain/model/box_movement.dart'
-    as _i16;
+    as _i18;
 import 'package:zachranobed/features/foodboxes/presentation/screen/box_movement_detail_screen.dart'
     as _i2;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
@@ -23,32 +23,36 @@ import 'package:zachranobed/features/menu/presentation/contacts_screen.dart'
     as _i4;
 import 'package:zachranobed/features/menu/presentation/menu_screen.dart'
     as _i10;
+import 'package:zachranobed/features/offeredfood/domain/model/food_info.dart'
+    as _i20;
 import 'package:zachranobed/features/offeredfood/domain/model/offered_food.dart'
-    as _i17;
+    as _i19;
 import 'package:zachranobed/features/offeredfood/presentation/screens/donated_food_detail_screen.dart'
     as _i6;
 import 'package:zachranobed/ui/screens/change_password_screen.dart' as _i3;
 import 'package:zachranobed/ui/screens/forgot_password_screen.dart' as _i7;
 import 'package:zachranobed/ui/screens/home_screen.dart' as _i8;
-import 'package:zachranobed/ui/screens/offer_food_screen.dart' as _i11;
+import 'package:zachranobed/ui/screens/offer_food_detail_screen.dart' as _i11;
+import 'package:zachranobed/ui/screens/offer_food_overview_screen.dart' as _i12;
+import 'package:zachranobed/ui/screens/offer_food_screen.dart' as _i13;
 import 'package:zachranobed/ui/screens/order_shipping_of_boxes_screen.dart'
-    as _i12;
-import 'package:zachranobed/ui/screens/thank_you_screen.dart' as _i13;
+    as _i14;
+import 'package:zachranobed/ui/screens/thank_you_screen.dart' as _i15;
 
-abstract class $AppRouter extends _i14.RootStackRouter {
+abstract class $AppRouter extends _i16.RootStackRouter {
   $AppRouter({super.navigatorKey});
 
   @override
-  final Map<String, _i14.PageFactory> pagesMap = {
+  final Map<String, _i16.PageFactory> pagesMap = {
     AppTermsRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i1.AppTermsScreen(),
       );
     },
     BoxMovementDetailRoute.name: (routeData) {
       final args = routeData.argsAs<BoxMovementDetailRouteArgs>();
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: _i2.BoxMovementDetailScreen(
           key: args.key,
@@ -57,26 +61,26 @@ abstract class $AppRouter extends _i14.RootStackRouter {
       );
     },
     ChangePasswordRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i3.ChangePasswordScreen(),
       );
     },
     ContactsRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i4.ContactsScreen(),
       );
     },
     DebugRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i5.DebugScreen(),
       );
     },
     DonatedFoodDetailRoute.name: (routeData) {
       final args = routeData.argsAs<DonatedFoodDetailRouteArgs>();
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: _i6.DonatedFoodDetailScreen(
           key: args.key,
@@ -85,46 +89,67 @@ abstract class $AppRouter extends _i14.RootStackRouter {
       );
     },
     ForgotPasswordRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i7.ForgotPasswordScreen(),
       );
     },
     HomeRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i8.HomeScreen(),
       );
     },
     LoginRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i9.LoginScreen(),
       );
     },
     MenuRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i10.MenuScreen(),
       );
     },
-    OfferFoodRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+    OfferFoodDetailRoute.name: (routeData) {
+      final args = routeData.argsAs<OfferFoodDetailRouteArgs>();
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i11.OfferFoodScreen(),
+        child: _i11.OfferFoodDetailScreen(
+          key: args.key,
+          editedFoodInfo: args.editedFoodInfo,
+          allFoodInfos: args.allFoodInfos,
+        ),
+      );
+    },
+    OfferFoodOverviewRoute.name: (routeData) {
+      final args = routeData.argsAs<OfferFoodOverviewRouteArgs>();
+      return _i16.AutoRoutePage<dynamic>(
+        routeData: routeData,
+        child: _i12.OfferFoodOverviewScreen(
+          key: args.key,
+          foodInfos: args.foodInfos,
+        ),
+      );
+    },
+    OfferFoodRoute.name: (routeData) {
+      return _i16.AutoRoutePage<dynamic>(
+        routeData: routeData,
+        child: const _i13.OfferFoodScreen(),
       );
     },
     OrderShippingOfBoxesRoute.name: (routeData) {
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i12.OrderShippingOfBoxesScreen(),
+        child: const _i14.OrderShippingOfBoxesScreen(),
       );
     },
     ThankYouRoute.name: (routeData) {
       final args = routeData.argsAs<ThankYouRouteArgs>();
-      return _i14.AutoRoutePage<dynamic>(
+      return _i16.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: _i13.ThankYouScreen(
+        child: _i15.ThankYouScreen(
           key: args.key,
           isSuccess: args.isSuccess,
           message: args.message,
@@ -136,8 +161,8 @@ abstract class $AppRouter extends _i14.RootStackRouter {
 
 /// generated route for
 /// [_i1.AppTermsScreen]
-class AppTermsRoute extends _i14.PageRouteInfo<void> {
-  const AppTermsRoute({List<_i14.PageRouteInfo>? children})
+class AppTermsRoute extends _i16.PageRouteInfo<void> {
+  const AppTermsRoute({List<_i16.PageRouteInfo>? children})
       : super(
           AppTermsRoute.name,
           initialChildren: children,
@@ -145,17 +170,17 @@ class AppTermsRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'AppTermsRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i2.BoxMovementDetailScreen]
 class BoxMovementDetailRoute
-    extends _i14.PageRouteInfo<BoxMovementDetailRouteArgs> {
+    extends _i16.PageRouteInfo<BoxMovementDetailRouteArgs> {
   BoxMovementDetailRoute({
-    _i15.Key? key,
-    required _i16.BoxMovement boxMovement,
-    List<_i14.PageRouteInfo>? children,
+    _i17.Key? key,
+    required _i18.BoxMovement boxMovement,
+    List<_i16.PageRouteInfo>? children,
   }) : super(
           BoxMovementDetailRoute.name,
           args: BoxMovementDetailRouteArgs(
@@ -167,8 +192,8 @@ class BoxMovementDetailRoute
 
   static const String name = 'BoxMovementDetailRoute';
 
-  static const _i14.PageInfo<BoxMovementDetailRouteArgs> page =
-      _i14.PageInfo<BoxMovementDetailRouteArgs>(name);
+  static const _i16.PageInfo<BoxMovementDetailRouteArgs> page =
+      _i16.PageInfo<BoxMovementDetailRouteArgs>(name);
 }
 
 class BoxMovementDetailRouteArgs {
@@ -177,9 +202,9 @@ class BoxMovementDetailRouteArgs {
     required this.boxMovement,
   });
 
-  final _i15.Key? key;
+  final _i17.Key? key;
 
-  final _i16.BoxMovement boxMovement;
+  final _i18.BoxMovement boxMovement;
 
   @override
   String toString() {
@@ -189,8 +214,8 @@ class BoxMovementDetailRouteArgs {
 
 /// generated route for
 /// [_i3.ChangePasswordScreen]
-class ChangePasswordRoute extends _i14.PageRouteInfo<void> {
-  const ChangePasswordRoute({List<_i14.PageRouteInfo>? children})
+class ChangePasswordRoute extends _i16.PageRouteInfo<void> {
+  const ChangePasswordRoute({List<_i16.PageRouteInfo>? children})
       : super(
           ChangePasswordRoute.name,
           initialChildren: children,
@@ -198,13 +223,13 @@ class ChangePasswordRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'ChangePasswordRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i4.ContactsScreen]
-class ContactsRoute extends _i14.PageRouteInfo<void> {
-  const ContactsRoute({List<_i14.PageRouteInfo>? children})
+class ContactsRoute extends _i16.PageRouteInfo<void> {
+  const ContactsRoute({List<_i16.PageRouteInfo>? children})
       : super(
           ContactsRoute.name,
           initialChildren: children,
@@ -212,13 +237,13 @@ class ContactsRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'ContactsRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i5.DebugScreen]
-class DebugRoute extends _i14.PageRouteInfo<void> {
-  const DebugRoute({List<_i14.PageRouteInfo>? children})
+class DebugRoute extends _i16.PageRouteInfo<void> {
+  const DebugRoute({List<_i16.PageRouteInfo>? children})
       : super(
           DebugRoute.name,
           initialChildren: children,
@@ -226,17 +251,17 @@ class DebugRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'DebugRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i6.DonatedFoodDetailScreen]
 class DonatedFoodDetailRoute
-    extends _i14.PageRouteInfo<DonatedFoodDetailRouteArgs> {
+    extends _i16.PageRouteInfo<DonatedFoodDetailRouteArgs> {
   DonatedFoodDetailRoute({
-    _i15.Key? key,
-    required _i17.OfferedFood offeredFood,
-    List<_i14.PageRouteInfo>? children,
+    _i17.Key? key,
+    required _i19.OfferedFood offeredFood,
+    List<_i16.PageRouteInfo>? children,
   }) : super(
           DonatedFoodDetailRoute.name,
           args: DonatedFoodDetailRouteArgs(
@@ -248,8 +273,8 @@ class DonatedFoodDetailRoute
 
   static const String name = 'DonatedFoodDetailRoute';
 
-  static const _i14.PageInfo<DonatedFoodDetailRouteArgs> page =
-      _i14.PageInfo<DonatedFoodDetailRouteArgs>(name);
+  static const _i16.PageInfo<DonatedFoodDetailRouteArgs> page =
+      _i16.PageInfo<DonatedFoodDetailRouteArgs>(name);
 }
 
 class DonatedFoodDetailRouteArgs {
@@ -258,9 +283,9 @@ class DonatedFoodDetailRouteArgs {
     required this.offeredFood,
   });
 
-  final _i15.Key? key;
+  final _i17.Key? key;
 
-  final _i17.OfferedFood offeredFood;
+  final _i19.OfferedFood offeredFood;
 
   @override
   String toString() {
@@ -270,8 +295,8 @@ class DonatedFoodDetailRouteArgs {
 
 /// generated route for
 /// [_i7.ForgotPasswordScreen]
-class ForgotPasswordRoute extends _i14.PageRouteInfo<void> {
-  const ForgotPasswordRoute({List<_i14.PageRouteInfo>? children})
+class ForgotPasswordRoute extends _i16.PageRouteInfo<void> {
+  const ForgotPasswordRoute({List<_i16.PageRouteInfo>? children})
       : super(
           ForgotPasswordRoute.name,
           initialChildren: children,
@@ -279,13 +304,13 @@ class ForgotPasswordRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'ForgotPasswordRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i8.HomeScreen]
-class HomeRoute extends _i14.PageRouteInfo<void> {
-  const HomeRoute({List<_i14.PageRouteInfo>? children})
+class HomeRoute extends _i16.PageRouteInfo<void> {
+  const HomeRoute({List<_i16.PageRouteInfo>? children})
       : super(
           HomeRoute.name,
           initialChildren: children,
@@ -293,13 +318,13 @@ class HomeRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'HomeRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i9.LoginScreen]
-class LoginRoute extends _i14.PageRouteInfo<void> {
-  const LoginRoute({List<_i14.PageRouteInfo>? children})
+class LoginRoute extends _i16.PageRouteInfo<void> {
+  const LoginRoute({List<_i16.PageRouteInfo>? children})
       : super(
           LoginRoute.name,
           initialChildren: children,
@@ -307,13 +332,13 @@ class LoginRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'LoginRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i10.MenuScreen]
-class MenuRoute extends _i14.PageRouteInfo<void> {
-  const MenuRoute({List<_i14.PageRouteInfo>? children})
+class MenuRoute extends _i16.PageRouteInfo<void> {
+  const MenuRoute({List<_i16.PageRouteInfo>? children})
       : super(
           MenuRoute.name,
           initialChildren: children,
@@ -321,13 +346,96 @@ class MenuRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'MenuRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i11.OfferFoodScreen]
-class OfferFoodRoute extends _i14.PageRouteInfo<void> {
-  const OfferFoodRoute({List<_i14.PageRouteInfo>? children})
+/// [_i11.OfferFoodDetailScreen]
+class OfferFoodDetailRoute
+    extends _i16.PageRouteInfo<OfferFoodDetailRouteArgs> {
+  OfferFoodDetailRoute({
+    _i17.Key? key,
+    required _i20.FoodInfo editedFoodInfo,
+    required List<_i20.FoodInfo> allFoodInfos,
+    List<_i16.PageRouteInfo>? children,
+  }) : super(
+          OfferFoodDetailRoute.name,
+          args: OfferFoodDetailRouteArgs(
+            key: key,
+            editedFoodInfo: editedFoodInfo,
+            allFoodInfos: allFoodInfos,
+          ),
+          initialChildren: children,
+        );
+
+  static const String name = 'OfferFoodDetailRoute';
+
+  static const _i16.PageInfo<OfferFoodDetailRouteArgs> page =
+      _i16.PageInfo<OfferFoodDetailRouteArgs>(name);
+}
+
+class OfferFoodDetailRouteArgs {
+  const OfferFoodDetailRouteArgs({
+    this.key,
+    required this.editedFoodInfo,
+    required this.allFoodInfos,
+  });
+
+  final _i17.Key? key;
+
+  final _i20.FoodInfo editedFoodInfo;
+
+  final List<_i20.FoodInfo> allFoodInfos;
+
+  @override
+  String toString() {
+    return 'OfferFoodDetailRouteArgs{key: $key, editedFoodInfo: $editedFoodInfo, allFoodInfos: $allFoodInfos}';
+  }
+}
+
+/// generated route for
+/// [_i12.OfferFoodOverviewScreen]
+class OfferFoodOverviewRoute
+    extends _i16.PageRouteInfo<OfferFoodOverviewRouteArgs> {
+  OfferFoodOverviewRoute({
+    _i17.Key? key,
+    required List<_i20.FoodInfo> foodInfos,
+    List<_i16.PageRouteInfo>? children,
+  }) : super(
+          OfferFoodOverviewRoute.name,
+          args: OfferFoodOverviewRouteArgs(
+            key: key,
+            foodInfos: foodInfos,
+          ),
+          initialChildren: children,
+        );
+
+  static const String name = 'OfferFoodOverviewRoute';
+
+  static const _i16.PageInfo<OfferFoodOverviewRouteArgs> page =
+      _i16.PageInfo<OfferFoodOverviewRouteArgs>(name);
+}
+
+class OfferFoodOverviewRouteArgs {
+  const OfferFoodOverviewRouteArgs({
+    this.key,
+    required this.foodInfos,
+  });
+
+  final _i17.Key? key;
+
+  final List<_i20.FoodInfo> foodInfos;
+
+  @override
+  String toString() {
+    return 'OfferFoodOverviewRouteArgs{key: $key, foodInfos: $foodInfos}';
+  }
+}
+
+/// generated route for
+/// [_i13.OfferFoodScreen]
+class OfferFoodRoute extends _i16.PageRouteInfo<void> {
+  const OfferFoodRoute({List<_i16.PageRouteInfo>? children})
       : super(
           OfferFoodRoute.name,
           initialChildren: children,
@@ -335,13 +443,13 @@ class OfferFoodRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'OfferFoodRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i12.OrderShippingOfBoxesScreen]
-class OrderShippingOfBoxesRoute extends _i14.PageRouteInfo<void> {
-  const OrderShippingOfBoxesRoute({List<_i14.PageRouteInfo>? children})
+/// [_i14.OrderShippingOfBoxesScreen]
+class OrderShippingOfBoxesRoute extends _i16.PageRouteInfo<void> {
+  const OrderShippingOfBoxesRoute({List<_i16.PageRouteInfo>? children})
       : super(
           OrderShippingOfBoxesRoute.name,
           initialChildren: children,
@@ -349,17 +457,17 @@ class OrderShippingOfBoxesRoute extends _i14.PageRouteInfo<void> {
 
   static const String name = 'OrderShippingOfBoxesRoute';
 
-  static const _i14.PageInfo<void> page = _i14.PageInfo<void>(name);
+  static const _i16.PageInfo<void> page = _i16.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i13.ThankYouScreen]
-class ThankYouRoute extends _i14.PageRouteInfo<ThankYouRouteArgs> {
+/// [_i15.ThankYouScreen]
+class ThankYouRoute extends _i16.PageRouteInfo<ThankYouRouteArgs> {
   ThankYouRoute({
-    _i15.Key? key,
+    _i17.Key? key,
     required bool isSuccess,
     required String message,
-    List<_i14.PageRouteInfo>? children,
+    List<_i16.PageRouteInfo>? children,
   }) : super(
           ThankYouRoute.name,
           args: ThankYouRouteArgs(
@@ -372,8 +480,8 @@ class ThankYouRoute extends _i14.PageRouteInfo<ThankYouRouteArgs> {
 
   static const String name = 'ThankYouRoute';
 
-  static const _i14.PageInfo<ThankYouRouteArgs> page =
-      _i14.PageInfo<ThankYouRouteArgs>(name);
+  static const _i16.PageInfo<ThankYouRouteArgs> page =
+      _i16.PageInfo<ThankYouRouteArgs>(name);
 }
 
 class ThankYouRouteArgs {
@@ -383,7 +491,7 @@ class ThankYouRouteArgs {
     required this.message,
   });
 
-  final _i15.Key? key;
+  final _i17.Key? key;
 
   final bool isSuccess;